% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IShares-class.R
\name{download_etf_constituents}
\alias{download_etf_constituents}
\title{Download IShares constituents}
\usage{
download_etf_constituents(summary_data, download_csv = TRUE,
  url_fixed_number, region)
}
\arguments{
\item{summary_data}{summary data dataframe as returned by \link{parse_summary_data}}

\item{download_csv}{if \code{TRUE} the constituents csv files will be saved in
a local folder called `csv_files`}

\item{url_fixed_number}{a country-specific code found in the url}

\item{region}{IShares geographical region (e.g. US)}
}
\value{
a named-list. Names are the \code{localExchangeTicker} found in \code{summary_data}
   and the elements are the csv file found in each ETF web page in melted format.
   By melted format we mean as output of the function \link{readr::melt_csv}
}
\description{
downloads from the IShares website the data regarding etf constituents. Constituent data
is publish in each etf web page as a csv file and we read it directly from there.
In \code{summary_data} there's the link pointing to the page where the csv is
published
}
