% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IShares-class.R
\name{download_summary_data.IShares}
\alias{download_summary_data.IShares}
\title{Download IShares summary}
\usage{
\method{download_summary_data}{IShares}(obj)
}
\arguments{
\item{\code{IShares}}{object}
}
\value{
a list with the following fields
   \itemize{
      \item \code{columns} column names
      \item \code{data} the data
   }
}
\description{
Download IShares summary
}
