% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IShares-class.R
\name{download_summary_data}
\alias{download_summary_data}
\title{Download summary data}
\usage{
download_summary_data(summary_link, as = "parsed")
}
\arguments{
\item{summary_link}{link where the IShares summary data (in JSON format) is downloaded from.
This is the data summarizing all the ETF of the IShares provider}

\item{as}{desired type of output: \code{raw}, \code{text} or
\code{parsed}. \code{content} attempts to automatically figure out
which one is most appropriate, based on the content-type.}
}
\value{
the JSON data downloaded from \code{summary_link} is the format specified by
   \code{as}
}
\description{
download the JSON summary data from the IShares website. For instance, this data
is rendered at the following address \link{https://www.ishares.com/uk/individual/en/products/etf-investments}
}
