% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_functions.R
\name{get_from_mongo}
\alias{get_from_mongo}
\title{Read from mongo}
\usage{
get_from_mongo(ticker, aod)
}
\arguments{
\item{ticker}{etf ticker (local exchange ticker plus country code)}

\item{aod}{etf as-of-date}
}
\value{
a list with the following attributes:
\itemize{
   \item \code{aod}
   \item \code{ticker}
   \item \code{summary_data}
   \item \code{constituents}
}
}
\description{
Read from mongo
}
