% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IShares-class.R
\name{IShares}
\alias{IShares}
\title{IShares class constructor}
\usage{
IShares(summary_link)
}
\arguments{
\item{summary_link}{link where the IShares summary data is downloaded from.
This is the data summarizing all the ETF of the IShares provider}
}
\description{
IShares class constructor
}
