% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IShares_IT-class.R
\name{IShares_IT}
\alias{IShares_IT}
\title{IShares_IT class constructor}
\usage{
IShares_IT(summary_link, ..., class = character())
}
\arguments{
\item{summary_link}{link where the IShares summary data is downloaded from.
This is the data summarizing all the ETF of the IShares provider}
}
\description{
IShares_IT class constructor
}
