% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IShares_UK-class.R
\name{IShares_UK}
\alias{IShares_UK}
\title{IShares_IT class constructor}
\usage{
IShares_UK(summary_link, get_constituents = FALSE,
  download_constituents_csv = FALSE, melted_constituents_list = list(),
  constituents_list = list(), tickers_to_keep = character(), ...,
  class = character())
}
\arguments{
\item{summary_link}{link where the IShares summary data (in JSON format) is downloaded from.
This is the data summarizing all the ETF of the IShares provider}

\item{tickers_to_keep}{character vector containing the tickers to keep. One might
want to keep just a subset of tickers due to storage constraints}
}
\description{
IShares_IT class constructor
}
