% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IShares-class.R
\name{parse_summary_data}
\alias{parse_summary_data}
\title{Parse IShares list}
\usage{
parse_summary_data(res_parsed)
}
\arguments{
\item{res_parsed}{a list returned by the \link{httr::content} function. The list
ha the following attributes: \itemize{
\item \code{code}
\item \code{message}
\item \code{status}
\item \code{code}
}
The most important attribute is data which contains the actual IShares data.
In particular, it's a list with the following attributes: \itemize{
   \item \code{tableData} a list with two attributes: \code{columns} and
      \code{data}. The first one store the column names and the second the actual data
   \item \code{config}
}}
}
\description{
Parse IShares list
}
