% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IShares_US-class.R
\name{parse_template_1_US}
\alias{parse_template_1_US}
\title{Parse template 1 US data}
\usage{
parse_template_1_US(melted_data)
}
\arguments{
\item{melted_data}{constituent data in melted format}
}
\value{
a parsed tibble with constituent data
}
\description{
this function parses the constituents data which is in the template 1 format.
After extracting the column names, the character values (indeed the function
readr::melt_csv convert all the data to char) are trimmed for removing blanks.
The SEDOL colum is forced to be char
}
