% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IShares_US-class.R
\name{parse_template_2_US}
\alias{parse_template_2_US}
\title{Parse template 2 US data}
\usage{
parse_template_2_US(melted_data)
}
\arguments{
\item{melted_data}{constituent dataframe in melted format}
}
\value{
a parsed tibble with constituent data
}
\description{
this function find the two blocks of data and passes them to \link{parse_template_1_US}
}
