% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IShares-class.R
\name{rename_col}
\alias{rename_col}
\title{Parse colum names}
\usage{
rename_col(df)
}
\arguments{
\item{df}{a datframe}
}
\value{
a dataframe with parsed column names
}
\description{
Most of the column names contain special characters or spaces. This function
gives them an uniform formatting by removing spaces and special characters
}
