% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryResult-class.R
\name{scale_weight}
\alias{scale_weight}
\title{Scale weights}
\usage{
scale_weight(obj)
}
\arguments{
\item{obj}{\code{QueryResult} object}
}
\description{
The instrument portfolio weights are recomputed as the ones given are rounded
up to the second digits
}
