% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IShares-class.R
\name{to_csv.IShares}
\alias{to_csv.IShares}
\title{Export to csv format}
\usage{
\method{to_csv}{IShares}(obj, output_folder)
}
\arguments{
\item{obj}{\code{IShares} object or one of its subclasses}

\item{output_folder}{folder where the csv files are to be saved}
}
\description{
\code{to_csv} export to csv format the dataframe \code{summary_data} and each
dataframe in \code{constituents_list}. The name of constituents files is the
ticker plus the region (e.g. IVV_US.csv)
}
