% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IShares-class.R
\name{to_mongo_data_format}
\alias{to_mongo_data_format}
\title{Convert data in mongo format}
\usage{
to_mongo_data_format(obj)
}
\arguments{
\item{obj}{a \code{IShares} object or one of its subclasses}
}
\value{
a named list. Elements are list with the following fields
   \itemize{
      \item \code{aod} nav as of date
      \item \code{ticker} local exchange ticker plus region
      \item \code{data} a list with summary and constiruents data
   }
}
\description{
The data is saved in mongo using the aod and ticker as keys and the data is
a list where the first field is the row from summary_data and the second the
constituents datafarme
}
