% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_functions.R
\name{write_IShares_to_mongo}
\alias{write_IShares_to_mongo}
\title{Write to Mongodb}
\usage{
write_IShares_to_mongo(data)
}
\arguments{
\item{data}{a element of the list returned by \link{to_mongo_data_format}. The attributes
are: \itemize{
   \item \code{aod} nav as of date
   \item \code{ticker} local exchange ticker plus region
   \item \code{data} a list with summary and constiruents data
}}
}
\description{
Write to Mongodb
}
