% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IShares-class.R
\name{IShares}
\alias{IShares}
\title{IShares class constructor}
\usage{
IShares(summary_link, tickers_to_keep)
}
\arguments{
\item{summary_link}{link where the IShares summary data (in JSON format) is downloaded from.
This is the data summarizing all the ETF of the IShares provider}

\item{tickers_to_keep}{character vector containing the tickers to keep. One might
want to keep just a subset of tickers due to storage constraints}
}
\value{
a new \code{IShares} object
}
\description{
The \code{IShares} class is a "virtual" class in the sense that it's never
actually instantiated but used to give the interface to its subclasses.
For instace at the \code{IShares} class level we implement the getter
methods
}
