% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IShares_US-class.R
\name{IShares_US}
\alias{IShares_US}
\title{IShares_US class constructor}
\usage{
IShares_US(summary_link, ..., class = character())
}
\arguments{
\item{summary_link}{link where the IShares summary data is downloaded from.
This is the data summarizing all the ETF of the IShares provider}
}
\description{
IShares_US class constructor
}
